Refactor to use app events and tests
