+++ src/js/enums/AppEvents.js
@@ -1,0 +1,13 @@
+/**
+ *
+ * @desc List of enums used for the app options. It's in CommonJS-like
+ *       format as some of these are used in the electron side of the app
+ * @type {Object}
+ *
+ */
+/* eslint-disable no-undef */
+module.exports = {
+    QUIT:           '__app_quit',
+    FOCUSED:        '__app_focused',
+    MENU_TOGGLE:    '__app_menu_open',
+};
